be/requirements.txt
django == 3.2.16
django-rest-framework == 0.1.0  
mysqlclient == 2.1.1